python_api/requirements.txt
flask
flask-cors
torch
torchvision
timm
pillow
pytorch-grad-cam
numpy